feat: add files to MultiModalMessage
