+++ aidial_adapter_openai/dial_api/resource.py
@@ -7,7 +7,7 @@ from aidial_sdk.chat_completion import Attachment
 from pydantic import BaseModel, Field, root_validator
 
 from aidial_adapter_openai.dial_api.storage import FileStorage, download_file
-from aidial_adapter_openai.utils.resource import Resource
+from aidial_adapter_openai.utils.resource.base import Resource
 from aidial_adapter_openai.utils.text import truncate_string
