+++ aidial_adapter_openai/utils/image_tokenizer.py
@@ -11,7 +11,10 @@ from pydantic import BaseModel
 from aidial_adapter_openai.configuration.deployment_type import (
     ChatCompletionDeploymentType,
 )
-from aidial_adapter_openai.utils.image import ImageDetail, resolve_detail_level
+from aidial_adapter_openai.utils.resource.image import (
+    ImageDetail,
+    resolve_detail_level,
+)
 
 
 class ImageTokenizer(BaseModel):
