+++ aidial_adapter_openai/utils/multi_modal_message.py
@@ -4,8 +4,9 @@ from openai.types.chat import ChatCompletionContentPartTextParam
 from openai.types.chat.chat_completion_content_part_param import File
 from pydantic import BaseModel
 
-from aidial_adapter_openai.utils.image import ImageResource
-from aidial_adapter_openai.utils.resource import Resource
+from aidial_adapter_openai.utils.resource.base import Resource
+from aidial_adapter_openai.utils.resource.file import FileResource
+from aidial_adapter_openai.utils.resource.image import ImageResource
 
 
 def create_file_content_part(filename: str, resource: Resource) -> File:
@@ -23,5 +24,6 @@ def create_text_content_part(text: str) -> ChatCompletionContentPartTextParam:
 
 
 class MultiModalMessage(BaseModel):
-    images: List[ImageResource]
+    images: List[ImageResource] = []
+    files: List[FileResource] = []
     raw_message: dict
