+++ aidial_adapter_openai/utils/resource/base.py
