+++ aidial_adapter_openai/video_generation/azure/adapter.py
@@ -4,6 +4,7 @@ from typing import Any, Dict, List, assert_never
 
 import fastapi
 from aidial_sdk.chat_completion import Choice
+from aidial_sdk.chat_completion import Message as DIALMessage
 from aidial_sdk.chat_completion import Request as DIALRequest
 from aidial_sdk.chat_completion import Response as DIALResponse
 from aidial_sdk.chat_completion import Stage
@@ -71,15 +72,24 @@ def _get_configuration(request: dict) -> VideoGenerationConfig:
     return configuration
 
 
-def _get_prompt(request: DIALRequest) -> str:
+def _get_last_message(request: DIALRequest) -> DIALMessage:
     messages = request.messages
-    prompt = messages[-1].content
+    if len(messages) == 0:
+        raise RequestValidationError("Expected at least one message.")
+    return messages[-1]
+
+
+def _get_prompt(request: DIALRequest) -> str:
+    prompt = _get_last_message(request)
+
     if not isinstance(prompt, str):
         raise RequestValidationError(
             "The last message must contain a text content."
         )
+
     if not prompt.strip():
         raise RequestValidationError("The prompt cannot be empty.")
+
     return prompt
